Partially fix Issue #64
In pileup, sid was a uint8_t in the struct definition, meaning there
could only be 256 samples at most when processing. Additional samples
would be ignored. Changing this to an int matches the type seen
throughout pileup.c and allows 257+ samples to be processed.

+++ src/pileup.h
@@ -65,7 +65,7 @@ typedef struct {
 void pileup_conf_init(pileup_conf_t *conf);
 
 typedef struct {
-    uint8_t sid;        /* which sample */
+    int sid;            /* which sample */
     uint8_t bsstrand:1; /* bisulfite strand */
     uint8_t qual:7;     /* base quality */
     uint8_t strand:1;   /* read stand */
